move deviceType specific functions into device files
